Move .py files with the same functionalities as jupyter notebooks to analysis/scripts folder
