+++ analysis-scripts/NECRvsLength.py
@@ -1,3 +1,5 @@
+import sys
+sys.path.append('../analysis/')
 import SiemensQuadraProperties as sqp
 from ActivityTools import *
 from SimulationDataset import *
@@ -9,7 +11,8 @@ from multiprocessing import Pool
 import random
 import numpy as np
 import matplotlib.pyplot as mpl
-myColours=[]
+# myColours=[royalblue, darkorange, yellowgreen, hotpink, orangered]
+myColours = []
 for colour in mpl.cm.viridis( np.linspace( 0.05, 0.8, 4 ) ):
     myColours.append( colour )
 mpl.rcParams['axes.prop_cycle'] = mpl.cycler(color=myColours)
@@ -22,7 +25,7 @@ params = {'legend.fontsize': 15,
           'ytick.labelsize': 15}
 mpl.rcParams.update(params)
 
-def NECRatTimeF18( tracerData, crystalData, crystalActivity, detectorRadius, phantomLength, detectorMaterial, simulationWindow=1E-2, coincidenceWindow=4.7E-9 ):
+def NECRatTimeF18( tracerData, crystalData, crystalActivity, detectorRadius, phantomLength, detectorMaterial, simulationWindow=1E-3, coincidenceWindow=4.7E-9 ):
     # NEMU NU 2-2012 says 650mm window for 700mm phantom, so keep the same relation
     zWindow = (phantomLength - 50) / 2
 
@@ -36,7 +39,7 @@ def NECRatTimeF18( tracerData, crystalData, crystalActivity, detectorRadius, pha
     activityAtTime = []
     for time in range( 0, 700, 20 ):
         timeSec = float(time) * 60.0
-        activity = F18ActivityAtTime( 20E6, timeSec )
+        activity = F18ActivityAtTime( 1100E6, timeSec )
 
         activityList = []
         dataList = []
@@ -56,49 +59,12 @@ def NECRatTimeF18( tracerData, crystalData, crystalActivity, detectorRadius, pha
     mpl.clf()
     return activityAtTime, necrAtTime, trueAtTime, rPlusSAtTime
 
-def NECRatTimeZr89( tracerData, crystalData, crystalActivity, detectorRadius, phantomLength, detectorMaterial, simulationWindow=1E-2, coincidenceWindow=4.7E-9 ):
-    # NEMU NU 2-2012 says 650mm window for 700mm phantom, so keep the same relation
-    zWindow = (phantomLength - 50) / 2
-
-    # get volume in cc
-    phantomRadius = 20.3 / 2.0
-    phantomVolume = phantomRadius * phantomRadius * math.pi * phantomLength / 10.0 # assume length in mm
-
-    necrAtTime = []
-    trueAtTime = []
-    rPlusSAtTime = []
-    activityAtTime = []
-    for time in range( 0, 700, 20 ):
-        timeSec = float(time) * 60.0
-        activity = Zr89ActivityAtTime( 20E6, timeSec )
-
-        activityList = []
-        dataList = []
-        # Hanna: Add crystalActivity to the activityList only if the crystal material is radioactive
-        if detectorMaterial == "LSO" or detectorMaterial == "LYSO" :
-            activityList = [activity, crystalActivity]
-            dataList = [tracerData, crystalData]
-        else :
-            activityList = [activity]
-            dataList = [tracerData]
-        necr, true, rPlusS, scatters, randoms = DetectedCoincidences( activityList, dataList, simulationWindow, coincidenceWindow, detectorRadius, ZMin=-zWindow, ZMax=zWindow )
-        necrAtTime.append( necr )
-        trueAtTime.append( true )
-        rPlusSAtTime.append( rPlusS )
-        activityAtTime.append( activity / phantomVolume )
-
-    mpl.clf()
-    return activityAtTime, necrAtTime, trueAtTime, rPlusSAtTime
-
-def OneDetector( detectorLength, phantomLength, detectorMaterial, isotope, simulationWindow=1E-2, coincidenceWindow=4.7E-9 ):
-    tracerName = "Linear" + isotope
-    isotopeMethod = NECRatTimeF18
-    if isotope == "Zr89":
-        isotopeMethod = NECRatTimeZr89
+# Investigate ideal detector length with Siemens geometry
+def OneDetector( detectorLength, phantomLength, detectorMaterial, simulationWindow=1E-3, coincidenceWindow=4.7E-9 ):
     # Fix random seed for reproducibility, don't if you want variation
     # Has to be set in this method, not before, because this is where we enter the worker processes
     random.seed( detectorLength )
-    tracerData = CreateDataset( detectorLength, "Siemens", phantomLength, tracerName, 100000, 435.0, 585.0, detectorMaterial )
+    tracerData = CreateDataset( detectorLength, "Siemens", phantomLength, "LinearF18", 100000, 435.0, 585.0, detectorMaterial )
     crystalData = None
     crystalActivity = None
     #calculate crystal activity and create crystalDataset only if the crystal is radioactive
@@ -106,17 +72,17 @@ def OneDetector( detectorLength, phantomLength, detectorMaterial, isotope, simul
         crystalActivity= sqp.Lu176decaysInMass( sqp.DetectorMassLength( detectorLength, detectorMaterial ) )
         crystalData = CreateDataset( detectorLength, "Siemens", phantomLength, "Siemens", 100000, 435.0, 585.0, detectorMaterial )
 
-    activityAtTimeSiemens, necrAtTimeSiemens, trueAtTimeSiemens, rPlusSAtTimeSiemens = isotopeMethod( tracerData, crystalData, crystalActivity, sqp.DetectorRadius(), phantomLength, detectorMaterial, simulationWindow, coincidenceWindow )
+    activityAtTimeSiemens, necrAtTimeSiemens, trueAtTimeSiemens, rPlusSAtTimeSiemens = NECRatTimeF18( tracerData, crystalData, crystalActivity, sqp.DetectorRadius(), phantomLength, detectorMaterial, simulationWindow, coincidenceWindow )
     return ( max( necrAtTimeSiemens ), sqp.DetectorDiscreteLength( detectorLength ) )
 
-def PeakNECRWithLengthMultiprocess( phantomLength, detectorMaterial, isotope, simulationWindow=1E-2, coincidenceWindow=4.7E-9, processes=5 ):
+def PeakNECRWithLengthMultiprocess( phantomLength, detectorMaterial, simulationWindow=1E-3, coincidenceWindow=4.7E-9, processes=5 ):
     # Create the arguments for each process
     detectorLengths = [ 100, 300, 500, 700, 900, 1100, 1300, 1500, 1700, 1900 ]
     arguments = []
     realLengths = []
     maxNECR = []
     for detectorLength in detectorLengths:
-        arguments.append( ( detectorLength, phantomLength, detectorMaterial, isotope, simulationWindow, coincidenceWindow ) )
+        arguments.append( ( detectorLength, phantomLength, detectorMaterial, simulationWindow, coincidenceWindow ) )
         # result = OneDetector(detectorLength, phantomLength, detectorMaterial, simulationWindow, coincidenceWindow)
         # maxNECR.append(result[0])
         # realLengths.append(result[1])
@@ -135,29 +101,11 @@ def PeakNECRWithLengthMultiprocess( phantomLength, detectorMaterial, isotope, si
 
 detectorLengths = []
 maxNECRlines = []
-detectorMaterial = "LSO"
-isotopes = [ "F18", "Zr89" ]
-for isotope in isotopes:
-    detectorLengths, maxNECR = PeakNECRWithLengthMultiprocess( 700, detectorMaterial, isotope, simulationWindow=1E-2 )
-    maxNECRlines.append( maxNECR )
-
-mpl.plot( detectorLengths, maxNECRlines[0], label=isotope[0], color='r', linewidth=4.0 )
-mpl.plot( detectorLengths, maxNECRlines[1], label=isotope[1], color='b', linewidth=4.0 )
-
-mpl.xlabel( "Detector length [mm]" )
-mpl.ylabel( "Max NECR [/sec]" )
-mpl.legend( isotopes )
-mpl.gcf().set_size_inches( 10, 10 )
-mpl.savefig("F18vsZr89NECR.pdf")
-mpl.clf()
-
-detectorLengths = []
-maxNECRlines = []
 trialPhantoms = [ 300, 700, 1100, 1500 ]
+detectorMaterial = "LSO"
 for phantomLength in trialPhantoms:
-    detectorLengths, maxNECR = PeakNECRWithLengthMultiprocess( phantomLength, detectorMaterial, "Zr89", simulationWindow=1E-2 )
+    detectorLengths, maxNECR = PeakNECRWithLengthMultiprocess( phantomLength, detectorMaterial, simulationWindow=1E-3, coincidenceWindow=4.7E-9 )
     maxNECRlines.append( maxNECR )
-
 
 for i, phantomLength in enumerate( trialPhantoms ):
     mpl.plot( detectorLengths, maxNECRlines[i], label=phantomLength, linewidth=4.0 )
@@ -166,8 +114,7 @@ mpl.xlabel( "Detector length [mm]" )
 mpl.ylabel( "Max NECR [/sec]" )
 mpl.legend( trialPhantoms, title="Source length [mm]" )
 mpl.gcf().set_size_inches( 10, 10 )
-mpl.savefig("NECRvsLengthZr89.pdf")
-mpl.clf()
+mpl.savefig('NECRvsLength.pdf')
 
 def scale( y0, y1, x1 ):
     m = (y1 - y0) / (x1 - detectorLengths[0])
@@ -183,12 +130,12 @@ def plotThreeParts( xs, ys, i, label, colour ):
 
 legendEntries = []
 for i, trialPhantom in enumerate( trialPhantoms ):
-    
+
     # find the detector length index closest to the source length
     detectorLengthIndex = 0
     while detectorLengths[ detectorLengthIndex ] < trialPhantom:
         detectorLengthIndex += 1
-    
+
     theplot = plotThreeParts( detectorLengths, maxNECRlines[i], detectorLengthIndex, trialPhantoms[i], myColours[i] )
     legendEntries.append( theplot )
 
@@ -196,5 +143,5 @@ mpl.xlabel( "Detector length [mm]")
 mpl.ylabel( "Max NECR [/sec]")
 mpl.legend( legendEntries, trialPhantoms, title="Source length [mm]" )
 mpl.gcf().set_size_inches(10,10)
-mpl.ylim( [0.0, 1.5e5] )
-mpl.savefig("NECRvsLengthZr89_withLine.pdf")
+mpl.ylim( [0.0, 2.5e6] )
+mpl.savefig('NECRvsLength_wLine.pdf')
